refactor: move cmc, ledger methods from commands/ to lib/
