+++ src/dfx/src/commands/ledger/mod.rs
@@ -1,31 +1,13 @@
 use crate::lib::agent::create_agent_environment;
 use crate::lib::environment::Environment;
 use crate::lib::error::DfxResult;
-use crate::lib::ledger_types::{
-    AccountIdBlob, BlockHeight, Memo, NotifyCreateCanisterArg, NotifyCreateCanisterResult,
-    NotifyTopUpArg, NotifyTopUpResult, TimeStamp, TransferArgs, TransferError, TransferResult,
-    MAINNET_CYCLE_MINTER_CANISTER_ID, MAINNET_LEDGER_CANISTER_ID,
-};
-use crate::lib::nns_types::account_identifier::{AccountIdentifier, Subaccount};
 use crate::lib::nns_types::icpts::ICPTs;
 use crate::NetworkOpt;
 
-use anyhow::{anyhow, bail, Context};
-use backoff::backoff::Backoff;
-use backoff::ExponentialBackoff;
-use candid::Principal;
-use candid::{Decode, Encode};
+use anyhow::anyhow;
 use clap::Parser;
 use fn_error_context::context;
-use ic_agent::agent::{RejectCode, RejectResponse};
-use ic_agent::agent_error::HttpErrorPayload;
-use ic_agent::{Agent, AgentError};
-use std::time::{SystemTime, UNIX_EPOCH};
 use tokio::runtime::Runtime;
-
-const TRANSFER_METHOD: &str = "transfer";
-const NOTIFY_TOP_UP_METHOD: &str = "notify_top_up";
-const NOTIFY_CREATE_METHOD: &str = "notify_create_canister";
 
 mod account_id;
 mod balance;
@@ -96,163 +78,5 @@ fn get_icpts_from_args(
             Ok(amount.map_err(|err| anyhow!(err))?)
         }
         Some(amount) => Ok(amount),
-    }
-}
-
-#[context("Failed to transfer funds.")]
-pub async fn transfer(
-    agent: &Agent,
-    canister_id: &Principal,
-    memo: Memo,
-    amount: ICPTs,
-    fee: ICPTs,
-    from_subaccount: Option<Subaccount>,
-    to: AccountIdBlob,
-    created_at_time: Option<u64>,
-) -> DfxResult<BlockHeight> {
-    let timestamp_nanos = created_at_time.unwrap_or(
-        SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap()
-            .as_nanos() as u64,
-    );
-
-    let mut retry_policy = ExponentialBackoff::default();
-
-    let block_height: BlockHeight = loop {
-        match agent
-            .update(canister_id, TRANSFER_METHOD)
-            .with_arg(
-                Encode!(&TransferArgs {
-                    memo,
-                    amount,
-                    fee,
-                    from_subaccount,
-                    to,
-                    created_at_time: Some(TimeStamp { timestamp_nanos }),
-                })
-                .context("Failed to encode arguments.")?,
-            )
-            .call_and_wait()
-            .await
-        {
-            Ok(data) => {
-                let result = Decode!(&data, TransferResult)
-                    .context("Failed to decode transfer response.")?;
-                match result {
-                    Ok(block_height) => break block_height,
-                    Err(TransferError::TxDuplicate { duplicate_of }) => {
-                        println!("{}", TransferError::TxDuplicate { duplicate_of });
-                        break duplicate_of;
-                    }
-                    Err(transfer_err) => bail!(transfer_err),
-                }
-            }
-            Err(agent_err) if !retryable(&agent_err) => {
-                bail!(agent_err);
-            }
-            Err(agent_err) => match retry_policy.next_backoff() {
-                Some(duration) => {
-                    eprintln!("Waiting to retry after error: {:?}", &agent_err);
-                    tokio::time::sleep(duration).await;
-                    println!("Sending duplicate transaction");
-                }
-                None => bail!(agent_err),
-            },
-        }
-    };
-
-    println!("Transfer sent at block height {block_height}");
-
-    Ok(block_height)
-}
-
-pub async fn transfer_cmc(
-    agent: &Agent,
-    memo: Memo,
-    amount: ICPTs,
-    fee: ICPTs,
-    from_subaccount: Option<Subaccount>,
-    to_principal: Principal,
-    created_at_time: Option<u64>,
-) -> DfxResult<BlockHeight> {
-    let to_subaccount = Subaccount::from(&to_principal);
-    let to =
-        AccountIdentifier::new(MAINNET_CYCLE_MINTER_CANISTER_ID, Some(to_subaccount)).to_address();
-    transfer(
-        agent,
-        &MAINNET_LEDGER_CANISTER_ID,
-        memo,
-        amount,
-        fee,
-        from_subaccount,
-        to,
-        created_at_time,
-    )
-    .await
-}
-
-pub async fn notify_create(
-    agent: &Agent,
-    controller: Principal,
-    block_height: BlockHeight,
-    subnet_type: Option<String>,
-) -> DfxResult<NotifyCreateCanisterResult> {
-    let result = agent
-        .update(&MAINNET_CYCLE_MINTER_CANISTER_ID, NOTIFY_CREATE_METHOD)
-        .with_arg(
-            Encode!(&NotifyCreateCanisterArg {
-                block_index: block_height,
-                controller,
-                subnet_type,
-            })
-            .context("Failed to encode notify arguments.")?,
-        )
-        .call_and_wait()
-        .await
-        .context("Notify call failed.")?;
-    let result =
-        Decode!(&result, NotifyCreateCanisterResult).context("Failed to decode notify response")?;
-    Ok(result)
-}
-
-pub async fn notify_top_up(
-    agent: &Agent,
-    canister: Principal,
-    block_height: BlockHeight,
-) -> DfxResult<NotifyTopUpResult> {
-    let result = agent
-        .update(&MAINNET_CYCLE_MINTER_CANISTER_ID, NOTIFY_TOP_UP_METHOD)
-        .with_arg(
-            Encode!(&NotifyTopUpArg {
-                block_index: block_height,
-                canister_id: canister,
-            })
-            .context("Failed to encode notify arguments.")?,
-        )
-        .call_and_wait()
-        .await
-        .context("Notify call failed.")?;
-    let result = Decode!(&result, NotifyTopUpResult).context("Failed to decode notify response")?;
-    Ok(result)
-}
-
-fn retryable(agent_error: &AgentError) -> bool {
-    match agent_error {
-        AgentError::ReplicaError(RejectResponse {
-            reject_code: RejectCode::CanisterError,
-            reject_message,
-            ..
-        }) if reject_message.contains("is out of cycles") => false,
-        AgentError::HttpError(HttpErrorPayload {
-            status,
-            content_type: _,
-            content: _,
-        }) if *status == 403 => {
-            // sometimes out of cycles looks like this
-            // assume any 403(unauthorized) is not retryable
-            false
-        }
-        _ => true,
     }
 }
