+++ src/dfx/src/commands/ledger/notify/create_canister.rs
@@ -1,5 +1,5 @@
-use super::super::notify_create;
 use crate::lib::ledger_types::NotifyError;
+use crate::lib::operations::cmc::notify_create;
 use crate::lib::root_key::fetch_root_key_if_needed;
 use crate::lib::{environment::Environment, error::DfxResult};
