+++ src/dfx/src/lib/operations/mod.rs
@@ -1,2 +1,3 @@
 pub mod canister;
+pub mod cmc;
 pub mod ledger;
